# Small performance improvment