--- a/src/libraries/System.Private.CoreLib/src/System/Numerics/Matrix3x2.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Numerics/Matrix3x2.cs
@@ -108,12 +108,7 @@ namespace System.Numerics
         /// <returns>A translation matrix.</returns>
         public static Matrix3x2 CreateTranslation(Vector2 position)
         {
-            Matrix3x2 result;
-
-            result.M11 = 1.0f;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
-            result.M22 = 1.0f;
+            Matrix3x2 result = _identity;
 
             result.M31 = position.X;
             result.M32 = position.Y;
@@ -129,12 +124,7 @@ namespace System.Numerics
         /// <returns>A translation matrix.</returns>
         public static Matrix3x2 CreateTranslation(float xPosition, float yPosition)
         {
-            Matrix3x2 result;
-
-            result.M11 = 1.0f;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
-            result.M22 = 1.0f;
+            Matrix3x2 result = _identity;
 
             result.M31 = xPosition;
             result.M32 = yPosition;
@@ -150,14 +140,10 @@ namespace System.Numerics
         /// <returns>A scaling matrix.</returns>
         public static Matrix3x2 CreateScale(float xScale, float yScale)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             result.M11 = xScale;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
             result.M22 = yScale;
-            result.M31 = 0.0f;
-            result.M32 = 0.0f;
 
             return result;
         }
@@ -171,14 +157,12 @@ namespace System.Numerics
         /// <returns>A scaling matrix.</returns>
         public static Matrix3x2 CreateScale(float xScale, float yScale, Vector2 centerPoint)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             float tx = centerPoint.X * (1 - xScale);
             float ty = centerPoint.Y * (1 - yScale);
 
             result.M11 = xScale;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
             result.M22 = yScale;
             result.M31 = tx;
             result.M32 = ty;
@@ -193,14 +177,10 @@ namespace System.Numerics
         /// <returns>A scaling matrix.</returns>
         public static Matrix3x2 CreateScale(Vector2 scales)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             result.M11 = scales.X;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
             result.M22 = scales.Y;
-            result.M31 = 0.0f;
-            result.M32 = 0.0f;
 
             return result;
         }
@@ -213,14 +193,12 @@ namespace System.Numerics
         /// <returns>A scaling matrix.</returns>
         public static Matrix3x2 CreateScale(Vector2 scales, Vector2 centerPoint)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             float tx = centerPoint.X * (1 - scales.X);
             float ty = centerPoint.Y * (1 - scales.Y);
 
             result.M11 = scales.X;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
             result.M22 = scales.Y;
             result.M31 = tx;
             result.M32 = ty;
@@ -235,14 +213,10 @@ namespace System.Numerics
         /// <returns>A scaling matrix.</returns>
         public static Matrix3x2 CreateScale(float scale)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             result.M11 = scale;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
             result.M22 = scale;
-            result.M31 = 0.0f;
-            result.M32 = 0.0f;
 
             return result;
         }
@@ -255,14 +229,12 @@ namespace System.Numerics
         /// <returns>A scaling matrix.</returns>
         public static Matrix3x2 CreateScale(float scale, Vector2 centerPoint)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             float tx = centerPoint.X * (1 - scale);
             float ty = centerPoint.Y * (1 - scale);
 
             result.M11 = scale;
-            result.M12 = 0.0f;
-            result.M21 = 0.0f;
             result.M22 = scale;
             result.M31 = tx;
             result.M32 = ty;
@@ -278,17 +250,13 @@ namespace System.Numerics
         /// <returns>A skew matrix.</returns>
         public static Matrix3x2 CreateSkew(float radiansX, float radiansY)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             float xTan = MathF.Tan(radiansX);
             float yTan = MathF.Tan(radiansY);
 
-            result.M11 = 1.0f;
             result.M12 = yTan;
             result.M21 = xTan;
-            result.M22 = 1.0f;
-            result.M31 = 0.0f;
-            result.M32 = 0.0f;
 
             return result;
         }
@@ -302,7 +270,7 @@ namespace System.Numerics
         /// <returns>A skew matrix.</returns>
         public static Matrix3x2 CreateSkew(float radiansX, float radiansY, Vector2 centerPoint)
         {
-            Matrix3x2 result;
+            Matrix3x2 result = _identity;
 
             float xTan = MathF.Tan(radiansX);
             float yTan = MathF.Tan(radiansY);
@@ -310,10 +278,9 @@ namespace System.Numerics
             float tx = -centerPoint.Y * xTan;
             float ty = -centerPoint.X * yTan;
 
-            result.M11 = 1.0f;
             result.M12 = yTan;
             result.M21 = xTan;
-            result.M22 = 1.0f;
+
             result.M31 = tx;
             result.M32 = ty;
 
@@ -327,8 +294,6 @@ namespace System.Numerics
         /// <returns>A rotation matrix.</returns>
         public static Matrix3x2 CreateRotation(float radians)
         {
-            Matrix3x2 result;
-
             radians = MathF.IEEERemainder(radians, MathF.PI * 2);
 
             float c, s;
@@ -367,12 +332,11 @@ namespace System.Numerics
             // [  c  s ]
             // [ -s  c ]
             // [  0  0 ]
+            Matrix3x2 result = _identity;
             result.M11 = c;
             result.M12 = s;
             result.M21 = -s;
             result.M22 = c;
-            result.M31 = 0.0f;
-            result.M32 = 0.0f;
 
             return result;
         }
@@ -773,9 +737,9 @@ namespace System.Numerics
         /// <returns>True if the Object is equal to this matrix; False otherwise.</returns>
         public override readonly bool Equals(object? obj)
         {
-            if (obj is Matrix3x2)
+            if (obj is Matrix3x2 m)
             {
-                return Equals((Matrix3x2)obj);
+                return Equals(m);
             }
 
             return false;