In the methods I modified, the return value has fields in common with the already-declared `_identity` static field.
So, instead of assigning field by field, I assign `_identity` directly then modify the different values.

Benchmarks code:

```csharp
    [Benchmark]
    public Matrix4x4[] OldImpl()
    {
        var arr = new Matrix4x4[2];
        for (int i = 0; i < 2; i++)
            arr[i] = Matrix4x4.CreateTranslation(1 + i, 2 + i, 3 + i);
        return arr;
    }

    [Benchmark]
    public Matrix4x4[] NewImpl()
    {
        var arr = new Matrix4x4[2];
        for (int i = 0; i < 2; i++)
            arr[i] = Matrix4x4.CreateTranslationOptimized(1 + i, 2 + i, 3 + i);
        return arr;
    }
```

Benchmark results:

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18362.900 (1903/May2019Update/19H1)
Intel Core i5-8250U CPU 1.60GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.300
  [Host]     : .NET Core 3.1.4 (CoreCLR 4.700.20.20201, CoreFX 4.700.20.22101), X64 RyuJIT
  DefaultJob : .NET Core 3.1.4 (CoreCLR 4.700.20.20201, CoreFX 4.700.20.22101), X64 RyuJIT
```

|  Method |     Mean |    Error |   StdDev |   Median |
|-------- |---------:|---------:|---------:|---------:|
| OldImpl | 67.72 ns | 8.491 ns | 25.04 ns | 56.45 ns |
| NewImpl | 63.37 ns | 8.067 ns | 23.79 ns | 56.39 ns |